Split InvestPage to components
